+++ src/main/java/dev/hoyin1600p/vault_render_optimization/client/LevelStateCleanup.java
@@ -1,6 +1,7 @@
 package dev.hoyin1600p.vault_render_optimization.client;
 
 import dev.hoyin1600p.vault_render_optimization.VaultRenderOptimization;
+import dev.hoyin1600p.vault_render_optimization.client.lighting.DynamicLightEngine;
 import dev.hoyin1600p.vault_render_optimization.mixin.CreateAdditionEnergyNetworkManagerAccessor;
 import dev.hoyin1600p.vault_render_optimization.mixin.PowahCableNetAccessor;
 import net.minecraftforge.api.distmarker.Dist;
@@ -20,6 +21,8 @@ public final class LevelStateCleanup {
     @SubscribeEvent(priority = EventPriority.LOWEST)
     public static void onWorldUnload(WorldEvent.Unload event) {
         Object world = event.getWorld();
+
+        DynamicLightEngine.clearIfWorld(world);
 
         if (ModList.get().isLoaded("createaddition")) {
             removeWorld(CreateAdditionEnergyNetworkManagerAccessor.vaultRenderOptimization$getInstances(), world,